feat: Add function to find center node in a graph
The code changes add a new function `find_center` that takes a list of edges as input and returns the center node of the graph. The function checks if the first node of the first edge is equal to either the first or second node of the second edge. If it is, then the center node is the first node of the first edge. Otherwise, the center node is the second node of the first edge.

The commit message suggests adding the function to find the center node in a graph.